refactor: use stable-dirname from unstable-dirname
