+++ path/posix/unstable_dirname.ts
@@ -1,9 +1,7 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import { assertArg } from "../_common/dirname.ts";
-import { stripTrailingSeparators } from "../_common/strip_trailing_separators.ts";
-import { isPosixPathSeparator } from "./_util.ts";
+import { dirname as stableDirname } from "./dirname.ts";
 import { fromFileUrl } from "./from_file_url.ts";
 
 /**
@@ -28,35 +26,5 @@ export function dirname(path: string | URL): string {
   if (path instanceof URL) {
     path = fromFileUrl(path);
   }
-  assertArg(path);
-
-  let end = -1;
-  let matchedNonSeparator = false;
-
-  for (let i = path.length - 1; i >= 1; --i) {
-    if (isPosixPathSeparator(path.charCodeAt(i))) {
-      if (matchedNonSeparator) {
-        end = i;
-        break;
-      }
-    } else {
-      matchedNonSeparator = true;
-    }
-  }
-
-  // No matches. Fallback based on provided path:
-  //
-  // - leading slashes paths
-  //     "/foo" => "/"
-  //     "///foo" => "/"
-  // - no slash path
-  //     "foo" => "."
-  if (end === -1) {
-    return isPosixPathSeparator(path.charCodeAt(0)) ? "/" : ".";
-  }
-
-  return stripTrailingSeparators(
-    path.slice(0, end),
-    isPosixPathSeparator,
-  );
+  return stableDirname(path);
 }
